Conflicts:
	FriendNGo/app/build.gradle
	FriendNGo/app/src/main/java/com/friendngo/friendngo/friendngo/ActivityNotifications.java

+++ FriendNGo/app/src/main/java/com/friendngo/friendngo/friendngo/Request.java
@@ -80,7 +80,7 @@ public class Request extends AppCompatActivity {
                                     activity.getString("sender_home_city"),
                                     activity.getString("sender_home_nationality"),
                                     activity.getString("sender_points"),
-                                    activity.getLong("id")));
+                                    activity.getInt("id")));
                         } catch (JSONException e) {
                             Log.w("GET REQUEST FAIL1: ", e.getMessage().toString());
                         }
